Added virtualization option to canvas instead of cropColumns.
Virtualization can be one of four option, false, 'both', 'row', 'col'.  This property controls what rows and columns are virtualized.  Rows and vitualized if the value is row and columns are virtualized if the value is col.  If the value is both both are vitualized.  If the value is anything else virtualization is disabled.